+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Motions list
 - Dispatch queue list and Scheduled list on Dispatch tab
 - Virtual list
+- `Referendums list`, `External list`, `Proposals list` on `Democracy page`
 
 #### Fixed
 - Collected winning data on auction tab